+++ src/main/java/org/deepsymmetry/beatlink/MediaDetails.java
@@ -135,12 +135,16 @@ public class MediaDetails {
         }
         mediaType = type;
 
-        try {
-            final String charsetName = hostPlayer < 40? "UTF-16BE" : "UTF-32LE";
-            name = new String(packetCopy, 0x2c, 0x40, charsetName).trim();
-            creationDate = new String(packetCopy, 0x6c, 0x18, charsetName).trim();
-        } catch (UnsupportedEncodingException e) {
-            throw new IllegalStateException("Java no longer supports UTF-16BE encoding?!", e);
+        if (hostPlayer < 40) {  // Does not send media name or creation date.
+            name = "rekordbox mobile";
+            creationDate = "";
+        } else {
+            try {
+                name = new String(packetCopy, 0x2c, 0x40, "UTF-16BE").trim();
+                creationDate = new String(packetCopy, 0x6c, 0x18, "UTF-16BE").trim();
+            } catch (UnsupportedEncodingException e) {
+                throw new IllegalStateException("Java no longer supports UTF-16BE encoding?!", e);
+            }
         }
 
         trackCount = (int)Util.bytesToNumber(packetCopy, 0xa6, 2);